Fix tests to work with component models
Why:

- Have tests check correctness

This change addresses the need by:

- Removing TimescaleDB compression
- Fixing the tests
- Moving the models
- Updating the consumer
