+++ core/Pipfile
@@ -13,7 +13,7 @@ codecov = "*"
 rope = "*"
 pytest = "*"
 pytest-django = "*"
-pytest-asyncio = "*"
+pytest-asyncio = "~=0.14.0"
 
 [packages]
 django = "~=3.1.0"
