+++ core/farms/README.md
@@ -5,3 +5,18 @@ To test the websocket, from a browser console run:
       console.debug("WebSocket message received:", event);
     };
     socket.send('{"type": "tel", "hello": "there"}')
+
+To test a command via the REST API (v1/farms/controllers/<uuid:pk>/command/), POST a request:
+
+    {
+      "type": "cmd",
+      "peripheral": {
+        "add": [
+          {
+            "name": "led33",
+            "type": "LED",
+            "pin": 33
+          }
+        ]
+      }
+    }
