+++ core/farms/models/__init__.py
@@ -1,0 +1,5 @@
+from farms.models.site import *
+from farms.models.controller import *
+from farms.models.peripheral import *
+from farms.models.data_point import *
+from farms.models.models import *
