--- core/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = core.settings
-python_files = farms/tests/test_ws_api.py
